VisionServiceIoTEdge/modules/CameraCapture/requirements.txt
azure-iot-device~=2.0.0
azure-iothub-device-client
numpy
# opencv-contrib-python
requests
trollius
tornado==4.5.3
picamera[array]
pypng
pillow